fix(gates): split capture-stash tests from lower/tests.rs (file cap, over since #9549)
